Refactor Statistics Popup tables
